frontend/requirements.txt
streamlit>=1.31.0
pillow>=10.2.0
numpy>=1.26.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
